Separate Delegates from main class
