+++ Sources/Controllers/BarcodeScannerViewController.swift
@@ -1,40 +1,6 @@
 import UIKit
 import AVFoundation
 
-// MARK: - Delegates
-
-/// Delegate to handle the captured code.
-public protocol BarcodeScannerCodeDelegate: class {
-  func scanner(
-    _ controller: BarcodeScannerViewController,
-    didCaptureCode code: String,
-    type: String
-  )
-}
-
-/// Delegate to report errors.
-public protocol BarcodeScannerErrorDelegate: class {
-  func scanner(_ controller: BarcodeScannerViewController, didReceiveError error: BarcodeScannerError)
-}
-
-/// Delegate to dismiss barcode scanner when the close button has been pressed.
-public protocol BarcodeScannerDismissalDelegate: class {
-  func scannerDidDismiss(_ controller: BarcodeScannerViewController)
-}
-
-// MARK: - Error types
-
-public enum BarcodeScannerError: Error {
-  /// Error when something besides a MachineReadableCodeObject was detected. (Check AVMetadataObject.ObjectType documentation)
-  case nonMachineReadableCodeDetected
-  /// Error describing an unexpected/general error
-  case unexpected(Error)
-  /// Error when a MachineReadableCodeObject was detected but its metadata is unsupported
-  case unsupported
-}
-
-// MARK: - Controller
-
 /**
  Barcode scanner controller with 4 sates:
  - Scanning mode
